created personShuffle function with most verb exceptions, must finish numberShuffle to complete personShuffle for verb choir 82
